move favorite logic to channel model
