+++ app/channel/template.hbs
@@ -26,6 +26,11 @@
 		</div>
 	</section>
 	<section class="Section">
+		<ul>
+			<li>canEdit: {{model.canEdit}}</li>
+			<li>isFavorite: {{model.isFavorite}}</li>
+			<li><button class="Btn" {{action 'toggleFavorite'}}>toggleFavorite</button></li>
+		</ul>
 		{{outlet}}
 	</section>
 </div>
