+++ app/utils/toggle-object.js
@@ -1,0 +1,7 @@
+export default function(context, obj, condition) {
+	if (condition) {
+		context.removeObject(obj);
+	} else {
+		context.addObject(obj);
+	}
+}
